Add flags to unmap extra memory in the tracee
There's a bunch of stuff in the tracee's memory image that isn't
necessary after it has been set up. We can inspect its /proc/$PID/maps
and execute munmap on it to clean that up.

There's a gross workaround for when seccomp() is enabled which opens up
just the munmap() syscall. The only better way I can think of is to use
PTRACE_O_SUSPEND_SECCOMP, but that requires CAP_SYS_ADMIN, which I'd
like to avoid.