Move to Node.js
One big goal I have for asmase is a web interface. The web interface and
the CLI would ideally share a lot of code which isn't possible if the
CLI is written in Python. So, this is the beginning of another big
rewrite to use JavaScript for everything. Linking to libasmase is kind
of clowny. That needs to be fixed.