libasmase: close the munmap() seccomp loophole
Currently, when we set up seccomp, we need to allow munmap() because we
call it later. However, it's actually not as hard as I thought to set up
seccomp by hand. Doing so means we can do the munmap from the tracer
first as before, and then set up seccomp from the tracer afterwards,
meaning the exception for munmap is no longer needed.