libasmase: mmap memfd at a fixed address
One idea I've been considering for the future is being able to save and
restore tracee state from persistent storage. This won't work very well
if the state includes absolute addresses into the memfd mapping and it
gets restored on a tracee with the memfd mapped somewhere else.

This commit makes it so that the memfd is always mapped at the same
address. This actually simplifies a couple of things:
1. The tracee now bootstraps itself instead of having the tracer run
   code on it.
2. The arch-specific machine code doesn't have to be encoded as raw
   bytes.